+++ backend/src/main/java/cc/remer/timetrack/domain/timeoff/TimeOffType.java
@@ -3,6 +3,7 @@ package cc.remer.timetrack.domain.timeoff;
 public enum TimeOffType {
   VACATION,
   SICK,
+  CHILD_SICK,
   PERSONAL,
   PUBLIC_HOLIDAY
 }